refact: Change to the pattern indicated by the library
